+++ forge/src/main/java/net/mehvahdjukaar/heartstone/forge/HeartstoneForge.java
@@ -4,10 +4,9 @@ import net.mehvahdjukaar.heartstone.Heartstone;
 import net.mehvahdjukaar.heartstone.HeartstoneClient;
 import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 /**
  * Author: MehVahdJukaar
@@ -19,16 +18,16 @@ public class HeartstoneForge {
         Heartstone.commonInit();
 
         if (PlatHelper.getPhysicalSide().isClient()) {
-            HeartstoneClient.init();
-            HeartstoneClientImpl.init();
+            MinecraftForge.EVENT_BUS.register(HeartstoneForge.class);
         }
-        FMLJavaModLoadingContext.get().getModEventBus().register(this);
     }
-
 
     @SubscribeEvent
-    public void setup(final FMLCommonSetupEvent event) {
-        event.enqueueWork(Heartstone::commonSetup);
+    public static void onClientTick(TickEvent.ClientTickEvent event) {
+        if (event.phase == TickEvent.Phase.START) {
+            HeartstoneClient.onClientTick();
+        }
     }
+
 }